journal: escape non-utf8 messages
$ printf '<13>Sep 15 15:07:58 HOST: null\0here' | nc -w1 -u -U /run/systemd/journal/dev-log
$ journalctl -n1 -o short-unix --no-hostname
1528452686.141537 HOST[18159]: null\000here

This mostly fixes #4863.

The standard allows NUL-bytes and arbitrary crap to be embeded in messages. We forward
all the crap to syslog, but strip whitespace and escape invalid characters to generate
a printable message to store as the MESSAGE= field.

The common case of messages that don't need stripping or escaping is optimized.