udev: Add migration code for /run/udev/links
If udevd is restarted after applying the changes for /run/udev/links
from "udev: change structure in /run/udev/links to improve efficiency",
it may find the old data structures left over from a previous version. Deal
with this situation by detecting it and moving the target files to the new
places.