networkd: fix integer overflow in prefix assignment
When receiving a prefix with size /56, networkd can assing 256 /64
prefixes from that. That value causes the uint8 "n_prefixes" to
overflow. Though an extreme example, there could be someone using Prefix
Delegation with a /18 from ULA space, which would even overflow an
uint32, so uint64 would be the safe choice.